Make Win64 compile.
Review URL: http://codereview.chromium.org/7389008

git-svn-id: http://v8.googlecode.com/svn/branches/experimental@8770 ce2b1a6d-e550-0410-aec6-3dcde31c8c00

+++ gc/src/heap.h
@@ -1202,7 +1202,7 @@ class Heap {
            (PromotedSpaceSize() + PromotedExternalMemorySize());
   }
 
-  inline void LowerOldGenLimits(int bytes) {
+  inline void LowerOldGenLimits(intptr_t bytes) {
     old_gen_promotion_limit_ -= bytes;
     old_gen_allocation_limit_ -= bytes;
   }
